Extracts bulk delete handler from main server code
